requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.7.7
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.45.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
load-dotenv==0.1.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.7
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.0
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-dotenv==0.19.0
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
